Nest `Join` class inside of `EvaluationStrategy`.
